+++ src/types/component.d.ts
@@ -9,3 +9,16 @@ export type ComponentProps = {
 	component: PageComponent;
 	dataObject: ComponentDataObject;
 };
+
+export type ComponentProfileProps = {
+	componentTitle: string;
+	componentTitleLink: string;
+	showComponentTitle: boolean;
+};
+
+// Joined with ComponentProfile but should extend?
+export type ComponentPropsObject = {
+	// this should be agiven no?
+	variant: string;
+} & ComponentProfileProps &
+	UnknownObject;
